start splitting grey vest components to their own files
