+++ src/themes/greyVestControls/IconButton.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+let IconButton = ({ className, primary, ...props }) => (
+  <div
+    className={`gv-icon-button ${className || ''} ${primary ? 'primary' : ''}`}
+    {...props}
+  />
+)
+export default IconButton
