+++ src/themes/greyVestControls/Textarea.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import { Observer } from 'mobx-react'
+
+let Textarea = React.forwardRef((props, ref) => (
+  <Observer>
+    {() => <textarea className="gv-input" {...props} ref={ref} />}
+  </Observer>
+))
+Textarea.displayName = 'Textarea'
+
+export default Textarea
